fix: listing of shared things and users
